Add new columns: created_at and updated_at (#7)
* Resolved dialog close

* Update API endpoint in requests.http and add p5-wrapper/react dependency

* Add a test automata grid

* Refactor main.tsx: Remove unused imports and comments, optimize cache for queries

* Initiating logic of saving video

* Giving up of recording video from Media Recorder

* Revert "Giving up of recording video from Media Recorder"

This reverts commit 8c6b37f4e321b9f4fccd9f26c82779f82e3b5e33.

* MP4 video generated

* Improving the p5 rerendering

* Imrpove routes

* Improving routes and add About link

* Add grid dimension, number of ingredients and custom names of them

* Fix a bug on creating and removing new names fields

* Cancel button

* Implement fieldArray

* Fix elements position on New Simulation

* Adicionando nomes e cores dos ingredientes

* Adicionando rota POST na api para nova simulação

* Integrando com banco de dados postgreSQL

* Matriz de J, falta corrigir from B to B

* Download input json file and fixing J parameter issue

* Import JSON file

* Add Pb

* Refactor simulation and calculation components; remove unused code and update schemas

* Remove calculation router from main application

* Remove unused steps and update commands for API dockerfile

* Simplificando arquivos para imagem docker e da api

* Adicionando suporte a migrations

* Gravando simulações no banco e retornando-as para a ui

* Adicionando toast e incluindo deleção e criação de cópias de simulações

* Refactor simulation model and enhance form validation; update UI components for improved layout and functionality

* Implement update simulation functionality and refine delete endpoint; enhance UI components for improved user experience

* Refactor Dockerfile and update database migration scripts; implement database creation logic and adjust service configurations

* Update README and enhance database model; add created_at and updated_at fields, and include iterations and results columns

* Refactor codebase: remove unused imports, update requirements, and implement alert dialog component

* Improve error handling in simulation toast notifications to display specific error messages

* Refactor simulation model to use CURRENT_TIMESTAMP for created_at and updated_at fields; update related queries and schemas

---------

+++ ui/src/routes/simulations/index.tsx
@@ -1,5 +1,17 @@
 import { EditSimulation } from "@/components/editSimulation";
 import { NewSimulation } from "@/components/newSimulation";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogActionDestructive,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 import {
   Card,
@@ -54,7 +66,10 @@ function SimulationsList() {
     toast.promise(mutation, {
       loading: "Deleting simulation...",
       success: `Simulation ${simulation?.name} deleted!`,
-      error: "Error deleting simulation",
+      error: (error) => {
+        const message = error.response?.data?.detail;
+        return message ?? `Error deleting ${simulation?.name}`;
+      },
     });
   };
 
@@ -70,7 +85,10 @@ function SimulationsList() {
     toast.promise(mutation, {
       loading: `Creating a copy of "${name}"...`,
       success: `Created a copy of "${name}"!`,
-      error: `Error copying ${name}`,
+      error: (error) => {
+        const message = error.response?.data?.detail;
+        return message ?? `Error creating a copy of ${simulation.name}`;
+      },
     });
   };
 
@@ -90,7 +108,7 @@ function SimulationsList() {
           <SimulationsSkeleton />
         ) : (
           <ul className="grid grid-cols-3 xl:grid-cols-4 gap-4 my-2">
-            {data?.map((simulation) => (
+            {data?.sort((a, b) => a.updated_at < b.updated_at ? 1 : -1).map((simulation) => (
               <Card key={simulation.id}>
                 <CardHeader>
                   <div className="flex justify-between items-center">
@@ -107,19 +125,42 @@ function SimulationsList() {
                       <TooltipContent>Edit</TooltipContent>
                     </Tooltip>
                     <CardTitle className="text-sm">{simulation.name}</CardTitle>
-                    <Tooltip>
-                      <TooltipTrigger asChild>
-                        <Button
-                          onClick={() => onDelete(simulation.id)}
-                          size="icon"
-                          variant="outline"
-                          className="w-8 h-8"
-                        >
-                          <Trash2Icon className="m-2 text-red-600" />
-                        </Button>
-                      </TooltipTrigger>
-                      <TooltipContent>Delete</TooltipContent>
-                    </Tooltip>
+                    <AlertDialog>
+                      <Tooltip>
+                        <AlertDialogTrigger asChild>
+                          <div>
+                            <TooltipTrigger asChild>
+                              <Button
+                                size="icon"
+                                variant="outline"
+                                className="w-8 h-8"
+                              >
+                                <Trash2Icon className="m-2 text-red-600" />
+                              </Button>
+                            </TooltipTrigger>
+                          </div>
+                        </AlertDialogTrigger>
+                        <TooltipContent>Delete</TooltipContent>
+                      </Tooltip>
+                      <AlertDialogContent>
+                        <AlertDialogHeader>
+                          <AlertDialogTitle>
+                            Are you sure you want to delete {simulation.name}?
+                          </AlertDialogTitle>
+                          <AlertDialogDescription>
+                            This action cannot be undone.
+                          </AlertDialogDescription>
+                        </AlertDialogHeader>
+                        <AlertDialogFooter>
+                          <AlertDialogCancel>Cancel</AlertDialogCancel>
+                          <AlertDialogActionDestructive
+                            onClick={() => onDelete(simulation.id)}
+                          >
+                            Yes, Delete
+                          </AlertDialogActionDestructive>
+                        </AlertDialogFooter>
+                      </AlertDialogContent>
+                    </AlertDialog>
                   </div>
                   <CardDescription className="text-xs">
                     <span>
@@ -136,18 +177,46 @@ function SimulationsList() {
                           .flatMap((i) => i.name)
                           .join(", ")}
                       </span>
+                        <span className="block">
+                          Created at: {new Date(simulation.created_at + 'Z').toLocaleString()}
+                        </span>
+                        <span className="block">
+                          Last updated at: {new Date(simulation.updated_at + 'Z').toLocaleString()}
+                        </span>
                     </span>
                   </CardDescription>
                 </CardHeader>
                 <CardFooter className="gap-2">
                   <div className="w-1/2">
-                    <Button
-                      onClick={() => onCopy(simulation.id)}
-                      className="w-full h-8 text-xs"
-                      variant="outline"
-                    >
-                      Create a copy
-                    </Button>
+                    <AlertDialog>
+                      <AlertDialogTrigger asChild>
+                        <Button
+                          className="w-full h-8 text-xs"
+                          variant="outline"
+                        >
+                          Create a copy
+                        </Button>
+                      </AlertDialogTrigger>
+                      <AlertDialogContent>
+                        <AlertDialogHeader>
+                          <AlertDialogTitle>
+                            Do you wish to create a copy of {simulation.name}?
+                          </AlertDialogTitle>
+                          <AlertDialogDescription>
+                            This action will create a new simulation with the
+                            same configurations as {simulation.name}.
+                          </AlertDialogDescription>
+                        </AlertDialogHeader>
+                        <AlertDialogFooter>
+                          <AlertDialogCancel>Cancel</AlertDialogCancel>
+                          <AlertDialogAction
+                            onClick={() => onCopy(simulation.id)}
+                          >
+                            Create
+                          </AlertDialogAction>
+                        </AlertDialogFooter>
+                      </AlertDialogContent>
+                    </AlertDialog>
                   </div>
                   <Link to={`/simulations/${simulation.id}`} className="w-1/2">
                     <Button className="w-full h-8 text-xs">Open</Button>
